move separate pages from workflow list to pages folder
