+++ packages/frinx-workflow-ui/src/pages/event-listeners/event-listeners.tsx
@@ -26,9 +26,9 @@ import {
   useToast,
 } from '@chakra-ui/react';
 import FeatherIcon from 'feather-icons-react';
-import { Editor, useNotifications, callbackUtils, EListener } from '@frinx/shared/src';
-import Paginator from '../../../common/pagination';
-import { usePagination } from '../../../hooks/use-pagination-hook';
+import { Editor, useNotifications, callbackUtils, EListener } from 'packages/shared/src';
+import Paginator from '../../common/pagination';
+import { usePagination } from '../../hooks/use-pagination-hook';
 
 const EventListeners = () => {
   const [eventListeners, setEventListeners] = useState<EListener[]>([]);
