+++ packages/frinx-workflow-ui/src/pages/executed-workflow-list/executed-workflow-bulk-operations-block/executed-workflow-bulk-operations.tsx
