+++ packages/frinx-workflow-ui/src/pages/executed-workflow-list/executed-workflow-table/executed-workflow-status-labels.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Tag } from '@chakra-ui/react';
-import { ExecutedWorkflowStatus } from '../../../../__generated__/graphql';
+import { ExecutedWorkflowStatus } from '../../../__generated__/graphql';
 
 type Props = {
   status: ExecutedWorkflowStatus | 'UNKNOWN';
