+++ packages/frinx-workflow-ui/src/pages/executed-workflow-list/executed-workflow-table/executed-workflow-table.tsx
@@ -3,7 +3,7 @@ import { Table, Tbody, Box, Text } from '@chakra-ui/react';
 import ExecutedWorkflowFlatTableItem from './executed-workflow-flat-table/executed-workflow-flat-table-item';
 import ExecutedWorkflowTableHead from './executed-workflow-table-head';
 import ExecutedWorkflowHierarchicalTableItem from './executed-workflow-hierarchical-table/executed-workflow-hierarchical-table-item';
-import { ExecutedWorkflowStatus, ExecutedWorkflowsQuery } from '../../../../__generated__/graphql';
+import { ExecutedWorkflowStatus, ExecutedWorkflowsQuery } from '../../../__generated__/graphql';
 import { SortProperty } from '../executed-workflow-list';
 
 type Props = {
