+++ packages/frinx-workflow-ui/src/pages/poll-data/poll-data.tsx
@@ -12,13 +12,13 @@ import {
   Thead,
   Tr,
 } from '@chakra-ui/react';
-import { callbackUtils, Queue } from '@frinx/shared/src';
+import { callbackUtils, Queue } from 'packages/shared/src';
 import FeatherIcon from 'feather-icons-react';
 import { orderBy } from 'lodash';
 import moment from 'moment';
 import React, { useEffect, useState } from 'react';
-import Paginator from '../../../common/pagination';
-import { usePagination } from '../../../hooks/use-pagination-hook';
+import Paginator from '../../common/pagination';
+import { usePagination } from '../../hooks/use-pagination-hook';
 
 function filterBySearchKeyword(queue: Queue, keyword: string): boolean {
   const query = keyword.toUpperCase();
