+++ packages/frinx-workflow-ui/src/pages/workflow-definitions/workflow-actions.tsx
@@ -13,7 +13,7 @@ import {
 } from '@chakra-ui/react';
 import FeatherIcon from 'feather-icons-react';
 import { Link } from 'react-router-dom';
-import { ClientWorkflow, jsonParse } from '@frinx/shared/src';
+import { ClientWorkflow, jsonParse } from 'packages/shared/src';
 
 type Props = {
   workflow: ClientWorkflow;
