+++ packages/frinx-workflow-ui/src/pages/workflow-definitions/workflow-definitions-header.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { Box, Button, Flex, Icon, IconButton, Input, InputGroup, FormLabel } from '@chakra-ui/react';
 import FeatherIcon from 'feather-icons-react';
-import WfAutoComplete from '../../../common/wf-autocomplete';
+import WfAutoComplete from '../../common/wf-autocomplete';
 
 type Props = {
   allLabels: string[];
