+++ packages/frinx-workflow-ui/src/pages/workflow-definitions/workflow-definitions.tsx
@@ -1,11 +1,11 @@
 import { Box, Container, Text, Progress, useDisclosure } from '@chakra-ui/react';
-import Pagination from '@frinx/inventory-client/src/components/pagination'; // TODO: can we move this to shared components?
-import { jsonParse, ClientWorkflow, Task } from '@frinx/shared/src';
+import Pagination from 'packages/frinx-inventory-client/src/components/pagination'; // TODO: can we move this to shared components?
+import { jsonParse, ClientWorkflow, Task } from 'packages/shared/src';
 import { debounce } from 'lodash';
 import React, { useMemo, useState } from 'react';
 import { gql, useMutation, useQuery } from 'urql';
 import { useNotifications } from '@frinx/shared';
-import { usePagination } from '../../../hooks/use-graphql-pagination';
+import { usePagination } from '../../hooks/use-graphql-pagination';
 import {
   DeleteWorkflowMutation,
   DeleteWorkflowMutationVariables,
@@ -13,7 +13,7 @@ import {
   UpdateWorkflowMutationVariables,
   WorkflowLabelsQuery,
   WorkflowsQuery,
-} from '../../../__generated__/graphql';
+} from '../../__generated__/graphql';
 import WorkflowDefinitionsHeader from './workflow-definitions-header';
 import WorkflowDefinitionsModals from './workflow-definitions-modals';
 import WorkflowDefinitionsTable from './workflow-definitions-table';
